refactor(setbuilder): extract useAgentChat + ChatPanelBody from ChatSidebar (#402)
Split the agent-chat logic into a reusable useAgentChat hook and a
presentational ChatPanelBody (with a shared PersonaToggle) so the desktop
sidebar and the upcoming mobile overlay can share one implementation.
ChatSidebar is now a thin desktop shell; its existing test suite stays green
as the behavior-preservation guard.
